Reworked the sponsor page, added Makefile, deleted _site from git
The sponsor page has now platinum sponsors. Add the to _config.yml
to have the sponsors listed.

The Makefile gives some shortcuts to run jekyll quicker.
It is removed from _site with the change to the end of the
_config.yml

_site is the result of the compilation, it should not be
normally stored on git (not sure here, maybe it is necessary
in the jekyll constellation of tools).